backend/requirements.txt
fastapi[standard]
uvicorn
sqlalchemy
pydantic
python-multipart
azure-identity
azure-mgmt-costmanagement
google-cloud-billing

